MODE-1189 Improve Large Property Support in Disk Connector
The attached patch adds a "large value area" to the disk connector.  Binary property values with a length that exceeds the new "largeValueSizeInBytes" field on the DiskSource will now be stored in a file in this area.  The SHA-1 hash of the value is Base64-encoded with an URL-safe encoding to generate a key for the value.  Each value is written to a file called <key>.dat in the large value area and the workspace and UUID of the owning node are stored in a file called <key>.ref in the large value area.

If the value is used in more than one place, only one data file will be created, but the reference file will be updated to add references to the other nodes that contain properties that use the value.

As properties are modified or deleted, their reference to the value is removed.  When the last reference to the value is removed, the data file and reference file for the value are deleted.

The location of the large values area defaults to a folder named "largeValues" under the repository root.  This location can be changed by setting the "largeValuePath" property on the DiskSource.

All tests pass.