Disk Connector with File Lock
The existing disk connector implementation uses an in-memory lock to guarantee a consistent state in the face of concurrent reads and writes.  This doesn't work for a clustered deployment, of course.

I've taken a crack at adding an (optional) lock approach that backs the in-memory lock with a file system lock on a lock file.  My thinking is that the lock file could be used to coordinate lock state among multiple JVMs within a cluster.

This patch is not ready for release, as it's completely undocumented, but if you get a few minutes, would you mind taking a look at the approach, particularly the code in DiskRepository.Disk*Lock?  I'd welcome any feedback.  Thanks!