MODE-1203 Support Node Caching in Connectors
The attached patch generalizes the earlier work towards a cache for the disk-based connector into a generic cache framework that can be used for the DiskSource, FileSystemSource, JdbcMetadataSource, and the SvnRepositorySource.  The InMemoryRepositorySource, InfinispanSource, and JBossCacheSource were not modified because they already store data in-memory and aren't likely to benefit from a node cache.  The JpaSource was not modified because it still uses the old map connector framework.  A decision should be made on whether to port that source forward into the new connector framework before caching is added.

A working cache implementation, InMemoryNodeCache is provided along with two working cache policies: InMemoryNodeCache and InMemoryNodeCache.  Connectors can use this cache by setting the nodeCachePolicy property on the connector to the classname of one of these implementations and setting the timeToLive property on the _cachePolicy_ (not on the source itself) to the length of time in seconds that nodes should remain in the cache.  Nodes are added to the cache on reads and writes.  Cache configurations have been added to the Reference Guide for each of the affected connectors.

IMPORTANT: THE CACHE POLICY MUST CORRESPOND TO THE CONNECTOR TYPE.  That is, path connectors must use a cache policy that creates caches that implement PathNodeCache, and map connectors must use a cache policy that creates caches that implement NodeCache.

All tests pass.