MODE-1201 Using InMemoryBinary when writing to a repository can cause heap/memory problems writing large files
Attached a very crude patch that creates a new kind of Binary implementation that will lazily wrap a FileInputStream to avoid trying to load it into memory and a test case that duplicates the issue.  The patch has some problems.  It is hacked into JcrValueFactory, it violates the immutability contract that Binary implementations are supposed to implement, and it only works for FileInputStreams.  The last issue may not be a real problem, as I can't think of another InputStream at this moment that wouldn't also be subject to heapsize limitations.

Nonetheless, it works around this issue, and this issue is going to pretty important for ModeShape when it's used as an enterprise repository's backing store.