MODE-1209 Add Support for Hibernate 2nd Level Cache to JPA Connector
Hard-coding a 2LC provider is really easy.  Making it usable is really hard, as Hibernate requires one to provide a CacheConcurrencyStrategy (e.g., read-write, read-only, transactional) and not all cache providers support all strategies.  The CacheConcurrencyStrategy can either be provided in a Hibernate-specifc  annotation at compile-time or in persistence.xml - a much more JPA-friendly mechanism.

Ergo, the attached patch adds the use of a persistence.xml file to help control mappings and Hibernate cache settings for the first time.  That, in turn, required that we use a more JPA-standard way of building an EntityManagerFactory, which required moving a lot of code, but was pretty straightforward.

The idea is that we can now specify a number of different pre-defined cache configurations in our persistence.xml and allow users to pick one by modifying the new jpaPersistenceUnitName property on the JpaSource.  Users can now also override or extend our default configurations by providing their own persistence.xml in their own JAR.

In doing this, I extracted all of the Hibernate-specific code into an implementation of a new interface, JpaAdapter.  This brings us surprisingly close to being able to support other JPA providers with our JPA connector, but we're still using Hibernate-specific annotations to define indexes, and we have to have those indexes.

Feedback on this approach would be greatly appreciated.  If the approach is considered viable, all that remains is to predefine some cache configurations and update the documentation accordingly.