Incorrect datatypes in ModeShape JDBC driver
Changed the resultset related classes to return jdbc column type name, instead of the jcr name as the column type name.    Also, had to fix the logic to match the uppercase jcr name returned in the query results in order to make the match to the JcrType used to provide the jdbc type.