MODE-456 Create SequencerOutputMap.addProperty Method
Added SequencerOutputMap.addPropertyValues(Path, Name, Object...) that adds values to the named property at the given path if it exists, or creates the named property at the given path with the given values if the property did not already exist.  This method will cause a PathNotFoundException if no node exists at the given path.

Added new tests to verify behavior and did a quick cleanup on the Getting Started guide where a few paragraphs reviewed the custom security context that no longer exists in the sequencer example.  All tests pass.