MODE-1228 Corrected restore of copied nodes below restored node
The logic of restoring a versioned node was not properly handling nodes below the versioned node - it was ignoring the 'jcr:mixinTypes' and 'jcr:uuid' properties. This change corrects that behavior to properly set these properties.

Note that while testing a fix, a case involving storing 'mix:versionable' nodes was found to not really be addressed by the JCR 2.0 specification. After some investigation, the reference implementation appears to use 'nt:frozenNode' for all nodes underneath a Version. This definitely works around the issue, so the ModeShape behavior was reverted (from MODE-1172) back to doing the same thing. The code is backward compatible for old version histories using node types other than nt:frozenNode (other than the 'mix:referenceable' issue).

Several tests were added to verify the problematic scenarios, and after the change all unit and integration tests pass. Similar tests using the reference implementation were added to the 'modeshape-test-reference-impl' module in the sandbox.