MODE-1237 Corrected handling of empty results from subquery.
Corrected how the query execution handles putting the subquery results into the variables, so that even
an empty array of tuples is put into the variables. Without this, when executing and using the subquery results,
the execution logic gets a null for the subquery 'variable' and records an error rather than just using an empty
array.

All unit and integration tests pass with these changes.