MODE-1150 Upgraded to SLF4J 1.6.1
Changed the SLF4J dependency from 1.5.11 to 1.6.1, including all of the documentation and the various web and jboss related modules.

Note that 1.6.x has changed its API slightly, and this breaks some libraries. Therefore, 1.6.x has to be used for all SLF4J-related libraries; one critical issue proved to be the "org.slf4j:jcl-over-slf4j" artifact that apparently is no longer needed (or at least needs to be the same version, and because RESTEasy had an explicit dependency on the then there was some work in excluding it from the web- and jboss- related ModeShape modules.

All unit and integration tests pass with these changes.