MODE-1241 Corrected duplicate serialization of properties in JPA source
The JPA source was handling UpdatePropertyRequests by iterating through its properties and then removing or
updating the property on the JPA node. Each time a property was modified (removed or added or updated), all
of the nodes properties were serialized. This not only is WAY more work than needed, but it causes a
problem when a file backed Binary value is used (since the first serialization closes the backing stream,
and subsequent serialization attempts to use the closed stream).

This was fixed by simply changing how the UpdatePropertyRequests are handled in MapRequestProcessor to
update and remove all of the appropriate properties in one shot, via a modified 'setProperties' method on
the MapNode, DefaultMapNode, and JpaNode. Because the MapNode interface was changed, several other uses
required modification.

Van's test case was able to duplicate the problem prior to these changes. After these changes, all
unit and integration tests pass.