MODE-1243 JndiRepositoryFactory should support registering engine in JNDI
The JndiRepositoryFactory implementation should support registering the ModeShape engine into JNDI if the repository name is empty or not provided. This change would allow using the factory to configure Glassfish or Tomcat by registering the engine in JNDI so that JCR-2.0-style lookup are possible.

See also MODE-1242 for the change to make RepositoryFactory recognize JNDI URLs to Repository instances, not just ModeShape engine instances.

All unit and integration tests pass with this change.