MODE-1250 Upgrade Apache POI to 3.7
The MS Office sequencer is using a different value of Apache POI than is being inherited from the Tika library used in the Tika Text Extractor module. This change upgrade the sequencer to also use 3.7, adds several unit and integration tests, and also makes a number of fixes to the way the sequencer works.

All unit and integration tests pass with these changes.