MODE-1243 Improved JndiRepositoryFactory logic and logging
The JndiRepositoryFactory now uses the "type" value (as specified in the 'conf/context.xml' for Tomcat) as the major impetus for registering the ModeShape JcrEngine instance or a javax.jcr.Repository instance. If an engine is to be registered, then the existence of a 'repositoryName' is logged as a warning. If a repository is to be registered, then the 'repositoryName' value is expected and must match an existing repository; all other cases are logged as an error.