MODE-1255 Corrected 'restore' logic
The logic for restoring nodes was not correctly getting the 'jcr:primaryType' value when checking
if the node in version history had a primary type of 'nt:frozenNode': it was comparing the
'nt:frozenNode' Name constant to the Property object (rather than the first value of the Property).
Thus, the if-check never succeeded, and the node is restored with a primary type of 'nt:frozenNode'.

After the simple fix, all unit and integration tests pass.