MODE-1263 Corrected indexing logic for file system content
Added the integration test supplied by Danny (with some minor modifications) and updated the @Before method in the FileSystemRepositoryIntegrationTest, and verified the query does not include any of the files or folders already present on the file system.

After a bit of debugging, I was able to figure out what the problem is and why we're only seeing it for the file system connector.

The actual bug is in SearchEngineIndexer logic: when the subgraph comes back with just a single node (i.e., the subgraph only contains a single level), the loop at line 336 will exit without processing the children of the single node in the subgraph. The fix is to check for this condition and add the children to the locationsToRead list before going into the loop.

The problem only manifests itself with the FileSystemConnector because it is the only connector to override the 'absoluteMaximumDepthForBranches()' method to return 1. All other connectors return the default value, 4 (or some other number > 1), and therefore the subgraph contains multiple levels and the while loop is entered and the children processed.

After the fix to SearchEngineIndexer, all unit and integration tests pass, including the new integration test.

Thanks again to Danny C for identifying this issue and creating an easily reproducible test case!