MODE-1269 Methods to re-index workspace content are now public
The re-indexing methods were previously on JcrSession, which was a package-protected
class, and thus the re-indexing methods were not publicly accessible. This change
moves the methods to the org.modeshape.jcr.api.Workspace interface is ModeShape's
public API. Now there are synchronous and asynchronous methods to re-index the entire
workspace and to re-index a subgraph below some path to a maximum depth.

Additional tests were added to verify the methods work. All unit and integration
tests pass.