MODE-1273 Removed the Seam Security provider
The authentication provider that is aware of Seam Security was removed because of the complexity
that even this 'provided' dependency added. It's still extremely easy for users to implement
an authentication provider that ties into their own security system, and we don't need to do
this for Seam Security any more than we do other frameworks.