MODE-1281 Corrected file connector's extra properties storage
When storing extra properties, the file system connector incorrectly wrote an empty multi-valued property
without the required trailing line separator. One example is the "jcr:predecessors" property, which is empty
upon first checkin of a 'mix:versionable' node. This caused an exception when reading the property back in.

This change corrects the way such empty properties are written to the file, but also adds logic to the
method that parses the file's lines to handle reading in files that were written with older versions.

The original problem was replicated with a new unit test and a new integration test, and then the
reading and writing logic were both independently verified. The new tests now pass.

All unit and integration tests pass.