MODE-1289 Implemented versioning and import/export
Created the JcrVersionManager that contained most of the logic from the legacy 'modeshape-jcr' module.
Much of that logic could be moved as is, but any of it that directly used the graph API or batches
had to be changed to use the new CachedNode API. The latter should be significantly more efficient
than the former, since the former was forcing at least one refresh (and sometimes multiple refreshes)
of the version history, while the CachedNode API doesn't require any refreshes.

Added support for importing and exporting repository content using system and document views, as
defined by the specification. Migrated the unit tests for the import/export functionality.

Note that unit tests for the versioning have not yet been migrated, but will be in forthcoming commits.