MODE-1289 Additional tests, implementations and fixes
Added more implementation and unit tests, and corrected the behavior of a number of features,
including the names, segments, and paths of renamed and moved nodes.

Added the use of a "@Migrated" annotation in the legacy codebase to help track which test
classes/methods have been migrated to the new codebase.

Also changed the Log4J configuration during tests.