MODE-1309 Corrected query behavior when using residual properties as the join criteria
Found and fixed a couple of problems. First, aliases were not being treated properly,
so any projected column with an alias was not actually connected to the underlying
property in the Lucene document. Secondly, the exception reported above was due to an
unnecessary check, which failed when the column corresponded to a residual property
(since the residual properties are not in the schema).

Added a number of new unit tests to verify that such queries (including variations)
are handled correctly. For example, the new tests verify that the criteria can reference
the aliased columns, even though neither of the supplied queries use the aliases in the criteria.

All unit and integration tests pass with these changes.