MODE-1314 Corrected JPA 2nd-level cache configuration for Hiberante 3.3+
The "cacheProviderClassName" can now safely be set to a {{RegionFactory}} implementation class, and the JPA connector does the appropriate thing. Note that when using a version of Hibernate earlier than 3.3, the class name of the {{ProviderClass}} implementation should be used for the value.

Also, a new property was added to the JPA connector in ModeShape 2.7. The {{cacheManagerLookup}} connector property can now be used when configuring Infinispan 2nd level cache to specify where in JNDI the Infinispan cache manager can be found.

All unit and integration tests pass.