MODE-1346 Added BinaryStore implementation
Added a new BinaryStore framework. Each JcrRepository instance now uses a BinaryStore
implementation to store all (non-trivial) Binary values. Two BinaryStore implementations
have been provided, including a FileSystemBinaryStore that stores binaries based upon
their SHA-1 hash on the local file system, and a TransientBinaryStore that stores binaries
based upon their SHA-1 hash in a directory within Java's 'java.io.tmpdir' directory.
The TransientBinaryStore is used by default, since without configuring an Infinispan
cache the content is stored in-memory anyway.

Note that with this new system in place, creating a JCR Binary value immediately stores
the stream's content inside the BinaryStore (unless it is below the minimum threshold
for storage, in which case it is stored in-memory and with the nodes), even before
the Session is saved. The JCR Binary value stores the key where the content is stored
inside the BinaryStore (the key is the hexadecimal form of the content's SHA-1 hash).
When the session is saved, the Binary's key is actually stored within the node's properties.
Note the RepositoryCache does keep track of the number of times each Binary key is referenced
by the nodes and properties, and any Binary keys that are no longer used will be
fired through the RepositoryCache listener framework via a new UnusedBinaryValue event.
The BinaryStore receives these unused keys, and marks them as unused. The repository's
garbage collection process then calls to the BinaryStore to remove any unused keys
that of a minimum age (which is defined to be twice the time period of the garbage
collection period).

Quite a few test cases were added to verify the behavior, and everything looks to be
working properly.

All unit and integration tests pass.