MODE-1324 Added methods to obtain SHA-1 has for a Binary value
Made the org.modeshape.jcr.api.Binary interface no longer deprecated (we're not using it again in
ModeShape 3.x, too), and added two methods for obtaining the SHA-1 hash of the value's content.
One of the methods returns the byte[] form of the hash, while the other returns the hexadecimal
string form of the same hash.